requirements.txt
streamlit
groq
PyPDF2
openpyxl
python-dotenv
pandas